Refactored so that the machine is genuinely comparing sizes and weights of coins, so the Coin is just a convenience and not over-used

+++ src/main/groovy/com/anthonyclifton/vendingmachine/Coin.groovy
@@ -22,10 +22,10 @@ enum Coin {
     NICKEL(21.21, 5.0),
     PENNY(19.05, 2.5)
 
-    float sizeInMillimeters
-    float weightInGrams
+    BigDecimal sizeInMillimeters
+    BigDecimal weightInGrams
 
-    Coin(float sizeInMillimeters, float weightInGrams) {
+    Coin(BigDecimal sizeInMillimeters, BigDecimal weightInGrams) {
         this.sizeInMillimeters = sizeInMillimeters
         this.weightInGrams = weightInGrams
     }
